m&m's sorter/requirements.txt
opencv-python
numpy
python-dotenv
langchain
langchain-openai